Created a new file named rectangle.php. We are going to be using this file to calculate the area of a rectangle. Remember, the area of a rectangle is calculated by multiplying the width by the height.
Created a class named Rectangle that contains the properties $height and $width, a constructor to set the height and width on instantiation, a method named area() that returns the area based on the height and width.

Created a shapes_test.php file next to rectangle.php. In shapes_test.php make sure to require_once rectangle.php.

Tested your new class by creating an instance of Rectangle with various heights and widths. Calling the area method should correctly display the product of height and width.

Created a new file named square.php