Updated the Rectangle class from the previous lesson to contain private properties for height and width.
Tested the Rectangle and Square classes to ensure desired results for area and parameter are still output.

Updated the height and width properties of the Rectangle class to have a visibility of protected.

Changed the visibility back to private for both properties. Create the necessary getters and setters for the Rectangle and Square classes to work as desired with only private properties.